Added JavaDoc + code clean up
